Move Theano-dependent code in optimizers package
All modules in garage.optimizers except for minibatch_dataset were moved
into the garage.theano package.
The corresponding imports were modified as well.
