+++ garage/optimizers/__init__.py
@@ -1,7 +1,1 @@
-from garage.optimizers.conjugate_gradient_optimizer import ConjugateGradientOptimizer
-from garage.optimizers.conjugate_gradient_optimizer import FiniteDifferenceHvp
 from garage.optimizers.minibatch_dataset import BatchDataset
-from garage.optimizers.first_order_optimizer import FirstOrderOptimizer  # noqa: I100
-from garage.optimizers.hf import HfOptimizer
-from garage.optimizers.lbfgs_optimizer import LbfgsOptimizer
-from garage.optimizers.penalty_lbfgs_optimizer import PenaltyLbfgsOptimizer
