+++ garage/regressors/gaussian_conv_regressor.py
@@ -8,13 +8,13 @@ import theano.tensor as TT
 from garage.core import Serializable
 from garage.misc import logger
 from garage.misc.ext import iterate_minibatches_generic
-from garage.optimizers import LbfgsOptimizer
-from garage.optimizers import PenaltyLbfgsOptimizer
 from garage.theano.core import ConvNetwork
 from garage.theano.core import LasagnePowered
 from garage.theano.core import ParamLayer
 from garage.theano.distributions import DiagonalGaussian
 from garage.theano.misc.tensor_utils import compile_function
+from garage.theano.optimizers import LbfgsOptimizer
+from garage.theano.optimizers import PenaltyLbfgsOptimizer
 
 
 class GaussianConvRegressor(LasagnePowered):
