+++ garage/theano/algos/tnpg.py
@@ -1,5 +1,5 @@
-from garage.optimizers import ConjugateGradientOptimizer
 from garage.theano.algos import NPO
+from garage.theano.optimizers import ConjugateGradientOptimizer
 
 
 class TNPG(NPO):
