+++ garage/theano/optimizers/__init__.py
@@ -1,0 +1,10 @@
+from garage.theano.optimizers.conjugate_gradient_optimizer import \
+        ConjugateGradientOptimizer
+from garage.theano.optimizers.conjugate_gradient_optimizer import \
+        FiniteDifferenceHvp
+from garage.theano.optimizers.first_order_optimizer import \
+        FirstOrderOptimizer
+from garage.theano.optimizers.hf import HfOptimizer
+from garage.theano.optimizers.lbfgs_optimizer import LbfgsOptimizer
+from garage.theano.optimizers.penalty_lbfgs_optimizer import \
+        PenaltyLbfgsOptimizer
