+++ garage/theano/optimizers/conjugate_gradient_optimizer.py
