+++ garage/theano/optimizers/first_order_optimizer.py
